+++ src/Rib/Markup.hs
@@ -11,7 +11,6 @@
 module Rib.Markup
   ( -- * Type class
     IsMarkup (..),
-    Markup (..),
   )
 where
 
@@ -19,13 +18,6 @@ import Data.Aeson (Value)
 import Lucid (Html)
 import Named
 import Path
-import Text.MMark (MMark)
-import Text.Pandoc (Pandoc)
-
--- A light-weight markup document structure
-data Markup doc where
-  Markup_Pandoc :: Markup Pandoc
-  Markup_MMark :: Markup MMark
 
 -- | Class for denoting Markup representations.
 --